try to get process for all conn Previously only retrieve during matching with a PROCESS-NAME rule.
