+++ adapter/inbound/socket.go
@@ -18,5 +18,6 @@ func NewSocket(target socks5.Addr, conn net.Conn, source C.Type) *context.ConnCo
 		metadata.SrcPort = port
 	}
 
+	metadata.ReadProcessName()
 	return context.NewConnContext(conn, metadata)
 }
